Fixed lock bug.
Now same user can re-issue a different lock
without problem. This is due to a probably bug
in Windows explorer client, when it creates a new
file it issue a LOCK, PUT 0kb then DELETE the file
and then issue another LOCK on the resource without
removing the previous lock.