Removed unneeded bits
This commit removes uneeded bits of WebDAVServer.

The previous commit replaced all HttpListener* objects.  That change
means the library can no longer be used as a standalone service.  It is
now only used to pass an HttpRequestBase object, and to use
MethodHandlers to set the given HttpResponseBase object.  It used by an
MVC controller which manages the route, authentication, etc. of the
request, and calls into WebDavSharp.Server when needed.